Fixes bug: Resets attemptedRows
After insert/delete POIs we were not reseting the counter that preserves
the number of rows affected before counting for the new affected rows.
So if we had two consequative deletions (without any insertions
inbetween), the rows affected on the second deletion should be 0. The
bug lead the counter to be equal to the number of affectedRows, of the
previous deletion.