Implements Update POIs for admin
It is a slow procedure and a bit difficult to demonstrate. Since we are
using starting_pois.json, in order to insert our POIs, the file that
updates those POIs should also be named starting_pois.json. Since it
contains quite a few POIs and the update operation is REALLY slow
(because we are executing 1 update for the poi_table and 24*7 updates
for the popular_times_table), we recommend earasing all the POIs you are
not going to update before uploading the file.
An alternative is to delete all popular_times_table data and re-insert
them, but that will also delete data from other files aswell (and thus
we will be forced to only accept a certain filename as the input file).
At the same time this will make it impossible to keep track of the
updates an admin makes, in admin_changes_table.
Another alternative is to just ignore any changes on the
popular_times_table and thus having to update only one record.